pyproject.toml
[project]
name = "divvy-transport-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair[all]>=6.0.0",
    "dbt>=1.0.0.40.7",
    "dbt-databricks>=1.10.14",
    "geopandas>=1.1.2",
    "matplotlib>=3.10.7",
    "python-dotenv>=1.2.1",
]
